refactor tests and add tests for environment file
